+++ Changelog.md
@@ -1,5 +1,6 @@
 # Next
 
+- **Breaking Change** Renamed `verbose` in the NetworkLoggerPlugin to `isVerbose`.
 - `TargetType` can specify its `ParameterEncoding`. The Default is `URLEncoding`.
 - Added documentation for `TargetType` and associated data structures.
 - Re-add `MultiTarget` to project.